Rename all GDT files to use a consistent naming scheme
